+++ R/icd.data-package.R
@@ -1,0 +1,14 @@
+#' @name icd.data
+#' @aliases icd.data-package
+#' @concept ICD-9 ICD-10 ICD
+#' @author Jack O. Wasey \email{jack@@jackwasey.com}
+#' @keywords misc utilities datasets
+#' @importFrom utils read.csv read.fwf
+#' @importFrom stats setNames
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL